Fixed MAJOR problems with karma tests.
It seems that for the last few commits no unit tests were running
and **I didn't notice it**. The test configuration included the
'angular-scenarios' package which shadowed the 'describe' function (?).
When looking at the output it all seemed green, but I didn't notice that
there were 0 tests running.

**NEED TO PAY MORE ATTENTION**

Then I had bunch of fixes to make to update the tests. Most important is
the call for 'httpFetch' instead of calling $http.get directly. This
caused most of the errors. The good thing is that I could remove all the
$httpBackend mocking.